Run tests as pytest smoke tests and get coverage reporting (#12)
* First part of #9, might actually solve #10.

Rename the tests so they run as pytest scripts. Not proper *unit* tests
but useful as smoke tests. Also add coverage reporting to XML and the
codecov upload action. 🤞

* Obviously we need the badge.

* Rename the final functions so we get explicit run information.

* Comments to PR #12.

Following an in-person chat. Adds an assert on the size of the error.
Explicitly skip one that is large, lower the reference level of the
grid.

* Update src/dxss/solve_2d.py